Optimized scheduled fetch in AnimeUpdateScheduler
Implemented checks so that existing anime that haven't changed been in the api won't be parsed by the parseSingleMedia() method in the AnimeService.

Next Will begin to add necessary fields to the anime Object such as hasBeenWatched and a "type" field that will mark certain anime as movies, or OVAs etc.

Once all forseen fields are added and populated in the database I need to work out Genre and Rating logic.

Sidenote: The rate limit prevents further speed optimization for the scheduled update. The "updatedAt" field doesn't have any helpful associated methods in the api for further optimization. Unfortunately the daily scheduled update will take approximately 860 seconds (about 15 mins) when it runs.