Initial implementation of React Components into the project
Still need to work on adding functionality for determining if the user has watched a particular show or not.

Maybe instead of adding an is watched boolean to the Anime Class just add a watchlist field to the anime class. Need to explore the pros and cons of both methods.